Adds kwargs option to Cmip5File __init__ to allow modification of given fp.
This allows for creating a Cmip5File object based on an existing file path (fp) template, and modifying one or more of its attributes.
This is useful for replacing "day" with "yr" in the freq and mip_table attributes for producing a file of annual averages calculated from the original fp.