Adds command line utility script for filling in the time and time_bnds variables in an annual average NetCDF.
This is for the case where an annual average NetCDF file (likely generated by CDO) may have incorrectly written time and time_bnds.  It reads the existing time units attribute to guess the start year for writing the values, but also allows the user to specify it with the -y command line option.