Adds an attribute check for automatic cache miss
Memoizing the grid mask in geo.py depends on having a hashable key and
the choice of that key affects performance. The 'model_id' netcdf
attribute is a decent choice for a key component, however we shouldn't
necessarily assume that it exists in the file. If a file does *not* have
the model_id attribute, this patch assures an automatic cache miss,
which prevents the code from failing altogether.

We still should follow through with issue #11 at some point (which will
obsolete this patch).