Uses GDAL subdataset notation to ensure proj info
For some NetCDF files with multiple variables, GDAL won't properly
recognize the projection information unless you also specify the
"subdataset" with the notation NETCDF:"[filename]":variable. Without
the projection information, it defaults to outputting a 512x512 image
which then won't match up with our mask at all.

This patch adds the use of the subdataset notation and also adds a check
for if GDAL falls back to its default indentity affine transformation
for the projection. If the check fails, the code raises a descriptive
Exception.