Push type checking for time param above utils
Lee's PR had added some type checking to the "time" parameter that is in
several API calls (stats, multistats, data, timeseries). The type
checking was buried at the bottom of the call stack in the utils module.
I advised that it should be done first thing (as soon as the API call is
made) and this patch implements that fix.