Move GDAL build steps into script for local use
Now that we want to run the leading edge of GDAL versions (2.3) it's a
pain in the @$$ to use in development (Ubuntu still doesn't have 2.3
in its base distro). Move the build steps that we use in TravisCI into
a script that we can easily invoke from our workstations in development.