Upgrades Sphinx and use system image Python
The base image for CI comes with GDAL and NetCDF Python modules
pre-installed. We want to use the system Python so that we don't have
to reinstall these (expensive to install) modules.

Additionally this patch upgrades Sphinx to ensure that CI can install
itself as the "ce" package and a few other packages got upgraded in
the Pipfile as a side effect.